Move the managers to separate module
