+++ floyd/client/base.py
@@ -1,7 +1,7 @@
 import requests
 
 import floyd
-from floyd.config import AuthConfigManager
+from floyd.manager.auth_config import AuthConfigManager
 from floyd.exceptions import AuthenticationException, BadRequestException, NotFoundException, OverLimitException
 from floyd.log import logger as floyd_logger
